+++ test/fixtures/console.js
@@ -1,3 +1,1 @@
-/* global window */
-
 window.console.log('Pavane')